+++ kbe/src/lib/network/tcp_packet_sender.cpp
@@ -223,7 +223,7 @@ Reason TCPPacketSender::processFilterPacket(Channel* pChannel, Packet * pPacket)
 		// Completion backends own a copy of the packet until WSASend completes, so the packet can leave the Channel queue now.
 		// 完成模型会在 WSASend 完成前持有 packet 副本，因此此处可以立即移出 Channel 队列。
 		const int remaining = pPacket->length() - pPacket->sentSize;
-		if (!pPoller->queueTcpSend(static_cast<int>(*pEndpoint), pPacket->data() + pPacket->sentSize, remaining))
+		if (!pPoller->queueTcpSend(*pEndpoint, pPacket->data() + pPacket->sentSize, remaining))
 		{
 			// 队列上限会设置 would-block 并保持可重试；无效或已关闭 socket 必须沿用同步发送的错误分类并关闭 Channel。
 			// Queue backpressure sets would-block and remains retryable; an invalid or closed socket follows synchronous-send error mapping and closes the Channel.
